+++ src/components/NavigationBar.js
@@ -5,7 +5,6 @@ export function NavigationBar() {
         <div className="NavigationBar-component">
             <div className="NavigationBar-links">
                 <Link to="/">Home</Link>
-                <Link to="/new-home">New Home</Link>
                 <Link to="/balances">Balances</Link>
             </div>
         </div>
